// SDKs/googleapis.com/composer/v1/typescript/src/sdk/models/shared/softwareconfig.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



// SoftwareConfig
/** 
 * Specifies the selection and configuration of software inside the environment.
**/
export class SoftwareConfig extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=airflowConfigOverrides" })
  airflowConfigOverrides?: Map<string, string>;

  @SpeakeasyMetadata({ data: "json, name=envVariables" })
  envVariables?: Map<string, string>;

  @SpeakeasyMetadata({ data: "json, name=imageVersion" })
  imageVersion?: string;

  @SpeakeasyMetadata({ data: "json, name=pypiPackages" })
  pypiPackages?: Map<string, string>;

  @SpeakeasyMetadata({ data: "json, name=pythonVersion" })
  pythonVersion?: string;

  @SpeakeasyMetadata({ data: "json, name=schedulerCount" })
  schedulerCount?: number;
}